requirements.txt
openai-whisper==20231117
torch==2.2.0
streamlit==1.32.0